cardano_vending_machine.py: Break apart NftVendingMachine class
The script is meant to be a lightweight invocation of the vending
machine using argparse and custom output directories.  To accomplish
this, we break apart an NftVendingMachine class into its own module with
a single ``vend()`` method.

The method's algorithm is:
1. Search for new UTXOs available in the payment address
2. Find how many mints are available in the vending machine
3. Do maths to see how many to mint and what to send to profit address
4. Perform the mint (manually to ensure txn balances)

Note that during testing we discovered some brittleness in the protocol
params file so we add a flag to pass in the ``protocol.json`` file path.

[ Documentation: None yet (PyDoc/README soon) ]
[ Testing: build, deploy to cluster, venv install, run script ]