NftVendingMachine: Support random vending
Many creators will generate limited edition artwork at the very end, but
want it to be vended randomly throughout their mint.  This provides a
"vend randomly" option that will shuffle the available mints when
os.listdir is called.  We seed the random in the demo script, however,
so that artists and project devs debugging this can reproduce any errors
that might occur.

NOTE: Per the [Python Documentation on
os.listdir](https://docs.python.org/3/library/os.html#os.listdir) the
files are already in an arbitrary order.  However, some OS'es may
default to sorting-like behavior, so we still think this is a useful
feature.

[ Documentation: README.md, help text ]
[ Testing: Manual to ensure reproducibility and randomization ]