Vending Machine: Support for free mints
We add the ability to support free mints here in ./main.py and update
the documentation.  This feature enables creators to take large
collections, ask the community for small amounts of ADA (two-dollar
bills) and return all the change they have plus a free NFT.  Lots of
folks are innovating in this space, so it is important to support them.

[ Documentation: README.md ]
[ Testing: Manual on testnet ]