Testing: Add simple smoketest for vending machine initialization
This is very simple: get the vending machine live, running without
error, and going to the preprod or preview blockchain to ensure that no
assets were minted.

NOTE: We support preprod AND preview given the split-brain Alonzo/Vasil
coding we are doing to ensure that both can be supported.

NOTE: Code coverage is now up to 49%

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv ]