cardano-cli: Remove deprecated cardano-node elements
This library is meant to run without any usage of the network.  Calls to
`cardano-cli` are merely in the codebase for creating wallets,
transactions, etc.  The deprecated code we removed here was creating
branches that can cause major runtime issues if we are not careful.
Remove all of them, and run all the tests to ensure no regressions.

[ Documentation: Updated README.md ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv -s ]