Vending Machine: Programatically generate the witness count i/o static
Before, we were automatically saying "Use this static witness count just
to be safe".  However, this could cost the vending machine operator
extra lovelace (maybe 30A-60A per 10K collection).  So, we change this
to be dynamically determined based on the number of transaction signers
as per:
https://developers.cardano.org/docs/integrate-cardano/multi-witness-transactions-cli/

NOTE: There are minor clean-ups and refactorings in the testing suite to
add a "burn" transaction directly to the smoke test thereby hardening
some of the pathways and additional args we could potentially pass to
the CardanoCli (and further testing our witness count usage).

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv -s ]