Vending Machine Core: Explicitly blacklist UTxOs with minUTxO issues
Sometimes the change or fee will be too little and it will cause us to
continually fail.  So, we simply raise a specific error to add to the
exclusions and avoid retrying.

NOTE: This creates a very small "leak" in our automated testing suite
because we do not have the ability to drain the payment UTxO at all and
it's locked in there.  There's no way to get around this at the current
moment.

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv -s ]