Vending Machine Core: Exile UTxOs with non-lovelace balances
Malicious or accidental actors may send native tokens or NFTs to the
minter and would cause an issue with minting due to an inability to
properly construct the remainder.  Rather than mess with all of the
potential corner cases for this, for now we explicitly exclude any
balances with non-zero non-lovelace balances.

NOTE: Total coverage is now at 94%

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv ]