Vending Machine Core: Fix bug with bad UTxO exclusion
We were incorectly excluding UTxOs that could be appropriately refunded
if the "final change" (user_rebate + change) were taken into account,
not just the change itself.  So, we change some of the logic in the code
and tighten up the padding in the test suite to trigger what used to be
this error, but now successfully passes.

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv -s ]