Testing: Initial implementation of manually run scale tests
Scale testing is not intended to be done regularly in the regression
suite (unless there are major changes where the behavior should be
confirmed to work with concurrent transactions).

To launch this test, you need to add flags to the pytest parameters.
For example:

        python3 -m coverage run --source=src/ --branch -m pytest \
                -k test_concurrent_wallet_usage \
                --num-wallets=3 \
                --mint-price=5000000 \
                --min-nfts=0 \
                --max-nfts=1 \
                --available-assets=2

In addition, if any error occurs during the test you can manually run
the `test_do_the_drain` test with an additional parameter labeled
`--old-test-dir=/path/to/prior/test` and it will manually burn/drain any
remaining UTxOs in the buyers for that test run.

NOTE: The min-nfts/max-nfts range is selected using a truncated normal
distribution not a uniform random selection.  This more accurately
models mint behavior where most users get "a couple".  Future
refinements could incorporate different types of skew.

[ Documentation: README.md (future commit) ]
[ Testing: Run over multiple iterations for correctness and speed ]