Whitelist: Fix bug related to zero_mints with valid whitelist
We anticipate the following potential "attack":
1. Actor finds out pay-to addy an hour before mint
2. Actor manually builds and submits “send to self” txn with wl asset
3. WL asset listed on secondary to get liquidity before mint
4. Vending machine starts up
5a. IF metadata is preloaded user who already sold WL asset would mint! [BAD]
5b. IF metadata EMPTY on startup Actor will get a refund automatically [GOOD]
     -> Vending machine restocked once all premature orders drained

Starting up empty was actually NOT supported due to a bug in the logic
of the SingleUseWhitelist `consume` function.  So we fix that and add
the test case below.

The test case ensures that refunding when the vending machine is empty
does NOT remove an asset from being properly whitelisted but also does
not create any assets.

[ Documentation: None ]
[ Testing: python3 -m coverage run --source=src/ --branch -m pytest -vv ]