Whitelist: Switch from txn input references to txn output references
Before, we were allowing subclasses of Whitelist to handle txn inputs of
UTxOs received by the vending machine in order to validate the place on
the whitelist.  This could be useful for example, if some custom
application wanted to redeem from a script at the same time as paying
the vending machine.

However, in the meantime it introduces a potential security flaw with
reference inputs.  That was already explicitly solved by the specific
implementations of Whitelist herein, but we wanted to remove branches
where possible in order to minimize the attack surface area.

This change switches whitelists to getting a list of txn *outputs* for
the txn that produced the currently-being-processed vending machine
payments.

[ Documentation: Updated pydoc ]
[ Testing: Full pytest suite ]