CLI: Enable "validate-only" mode for vending machine for just metadata
Before we run we do want to call `validate()` on all the constructed
objects but we actually want to start up the machine empty because it
will auto-refund anyone who leaked out the address and *then* move in
all the metadata.  This gets around the point of validation.

So, to fix this we create a subparser with options to exit the program
successfully after all the validation checks are performed.

[ Documentation: README.md, tests, and scripts ]
[ Testing: Manual CLI runs and pytest cli tests ]